Development (#2)
* Added CMake project to Ethereal

This CMake project will serve as the home for the Ethereal CAD kernel. The next step is to convert to a library and create a very simple test project to confirm this functionality.

* Trying to understand CMake I still have a long way to go.

Cmake seems really powerful but it's very foriegn to me as well. Using cmake will be super helpful in making this Ethereal run across OS's with minimal pain.

* Tried following cmake tutorial didn't work trying another one

* Cmake library working but needs polish

The camke library is finally working for me and I was able to link it to the simple point sample. The bad news is I had to past the project into the folder. This is a bad workflow so hopefully get_filename_component can help me fix it.

* Cmake is now working and building both the library and sample correctly

I finally found a way to make cmake automatically find the Ethereal library from the example code. This also greatly simplified and reduced the cmake script size.

Build system and still To Do:
-Add unit testing frame work.
-Documentation frame work.

* Revert "Cmake is now working and building both the library and sample correctly"

This reverts commit 3860505741597bd01ae645ce7b772435cbf6dcea.

* Revert "Revert "Cmake is now working and building both the library and sample correctly""

This reverts commit ed27dd71e76455f81610848dd6adf304efd39585.

* Cmake is actuallyworking now.

My past cmake setup was not correct. This one works and builds however it has several flaws. 1. The library path in simple point example (which still does not include any actual points, making it very simple) now points to the base directory library. However it uses a hard coded path which would only work on windows.
2. It currently only works as a ﬆatic lib.

Hopefully the next commit will able to ﬁx both of these issues.

* Updated to much improved cmake build system.

The new cmake build system provides much better building and linking. For the time being I am moving away from shared libraries as it should be easy enough to switch back and for the time being linking at compile time seems preferable for this project.