Fixed Google Tests
I'm not thrilled with the current results in VS test explorer namespace is empty this makes for a akward display but it's a good start.

Doesn't seem to be a good solution to this problem Link: https://developercommunity.visualstudio.com/content/problem/663974/test-explorer-namespace-seems-to-not-be-working.html

So I'll just deal for the time being.